refactor: on level portals and leveldata and textmeshpro added

+++ Assets/Scripts/GameManager.cs
@@ -4,7 +4,7 @@ namespace DefaultNamespace
 {
     public static class GameManager 
     {
-        public static Action<float> OnAnswerControl;
+        public static Action<int> OnAnswerControl;
         public static Action Victory;
         public static Action Failed;
     }
